agents/researcher/requirements.txt
strands-agents==1.19.0
bedrock-agentcore==1.1.1
boto3>=1.42.7
duckduckgo-search>=6.3.5
